fix: Extract claim button to make confetti work again
